+++ looper/const.py
@@ -81,6 +81,10 @@ __all__ = [
     "DEFAULT_CONFIG_FILEPATH",
     "DEFAULT_CONFIG_SCHEMA",
     "DEFAULT_COMPUTE_RESOURCES_NAME",
+    "MESSAGE_BY_SUBCOMMAND",
+    "SAMPLE_SELECTION_ATTRIBUTE_OPTNAME",
+    "SAMPLE_EXCLUSION_OPTNAME",
+    "SAMPLE_INCLUSION_OPTNAME",
 ]
 
 FLAGS = ["completed", "running", "failed", "waiting", "partial"]
@@ -228,3 +232,17 @@ PROFILE_COLNAMES = ["pid", "hash", "cid", "runtime", "mem", "cmd", "lock"]
 SAMPLE_SELECTION_ATTRIBUTE_OPTNAME = "sel-attr"
 SAMPLE_EXCLUSION_OPTNAME = "sel-excl"
 SAMPLE_INCLUSION_OPTNAME = "sel-incl"
+
+MESSAGE_BY_SUBCOMMAND = {
+    "run": "Run or submit sample jobs.",
+    "rerun": "Resubmit sample jobs with failed flags.",
+    "runp": "Run or submit project jobs.",
+    "table": "Write summary stats table for project samples.",
+    "report": "Create browsable HTML report of project results.",
+    "destroy": "Remove output files of the project.",
+    "check": "Check flag status of current runs.",
+    "clean": "Run clean scripts of already processed jobs.",
+    "inspect": "Print information about a project.",
+    "init": "Initialize looper config file.",
+    "init-piface": "Initialize generic pipeline interface.",
+}
