+++ tests/smoketests/test_cli_validation.py
@@ -4,14 +4,13 @@ import argparse
 from typing import *
 
 import pytest
-from looper import (
-    MESSAGE_BY_SUBCOMMAND,
+from looper.const import (
     SAMPLE_SELECTION_ATTRIBUTE_OPTNAME,
     SAMPLE_EXCLUSION_OPTNAME,
     SAMPLE_INCLUSION_OPTNAME,
 )
 from tests.conftest import print_standard_stream, subp_exec, test_args_expansion
-from looper.looper import main
+from looper.cli_looper import main
 
 
 SUBCOMMANDS_WHICH_SUPPORT_SKIP_XOR_LIMIT = ["run", "destroy"]
